% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_base_intermediaria_aneel.R
\name{cria_base_intermediaria_aneel}
\alias{cria_base_intermediaria_aneel}
\title{Cria a base intemediária para a aneel}
\usage{
cria_base_intermediaria_aneel(
 
    url_origem = "https://www.aneel.gov.br/documents/656831/16363897/SGPED_BI.zip/32560baa-fd7f-5a0d-3687-9085ae0dedc7"
)
}
\value{

}
\description{
Cria a base intemediária para a aneel
}
