% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_base_intermediaria_anp.R
\name{cria_base_intermediaria_anp}
\alias{cria_base_intermediaria_anp}
\title{Cria a base intemediária para a anp}
\usage{
cria_base_intermediaria_anp(
  origem_processos = here::here("data/ANP/projetos-rt-3-2015.csv")
)
}
\value{

}
\description{
Cria a base intemediária para a anp
}
