% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_base_intermediaria_bndes.R
\name{cria_base_intermediaria_bndes}
\alias{cria_base_intermediaria_bndes}
\title{Cria a base intemediária para o bndes criando um dataframe}
\usage{
cria_base_intermediaria_bndes(
  origem_processos = here::here("data/BNDES/naoautomaticas.xlsx")
)
}
\arguments{
\item{origem_processos}{dataset que contém os projetos de energia da fonte bndes}
}
\value{

}
\description{
Cria a base intemediária para o bndes criando um dataframe
}
\examples{
cria_base_intermediaria_bndes()
}
