% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_base_intermediaria_cnen.R
\name{cria_base_intermediaria_cnen}
\alias{cria_base_intermediaria_cnen}
\title{Cria a base intemediária para o CNEN criando um dataframe}
\usage{
cria_base_intermediaria_cnen(
  origem_processos = here::here("data/CNEN/Projeto CNEN_Plataforma Inova-E.xlsx")
)
}
\arguments{
\item{origem_processos}{dataset que contém os projetos da fonte cnen}
}
\value{

}
\description{
Cria a base intemediária para o CNEN criando um dataframe
}
\examples{
cria_base_intermediaria_cnen()
}
