% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_base_intermediaria_cnpq.R
\name{cria_base_intermediaria_cnpq}
\alias{cria_base_intermediaria_cnpq}
\title{Cria Base Intermediária CNPQ}
\usage{
cria_base_intermediaria_cnpq(
  origem_processos1 = here::here("data/CNPQ/Bolsas no Exterior 2004 a 2021.xlsx"),
  origem_processos2 = here::here("data/CNPQ/Bolsas Pais - 2004 a 2021.xlsx"),
  origem_processos3 = here::here("data/CNPQ/Fomento 2010-2021.xlsx")
)
}
\arguments{
\item{origem_processos2}{}
}
\value{

}
\description{
Cria Base Intermediária CNPQ
}
