% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_base_intermediaria_fapesp.R
\name{cria_base_intermediaria_fapesp}
\alias{cria_base_intermediaria_fapesp}
\title{Cria Base Intermediária da Fapesp}
\usage{
cria_base_intermediaria_fapesp(
 
    origem_processos = here::here("data/FAPESP/PROJETOS FAPESP SELECIONADOS INOVA-E - VALORES - 13 dez 2021.xlsx")
)
}
\arguments{
\item{origem_processos}{}
}
\value{

}
\description{
Cria Base Intermediária da Fapesp
}
