% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_base_intermediaria_finep.R
\name{cria_base_intermediaria_finep}
\alias{cria_base_intermediaria_finep}
\title{Cria a base intemediária para a finep  criando um dataframe}
\usage{
cria_base_intermediaria_finep(
  origem_processos = here::here("data/FINEP/14_09_2021_Liberacoes.ods")
)
}
\arguments{
\item{origem_processos}{dataset que contém os casos da fonte finep}
}
\value{

}
\description{
Cria a base intemediária para a finep  criando um dataframe
}
\examples{
cria_base_intermediaria_finep()
}
