% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_termos_gesi.R
\name{dtc_gesi}
\alias{dtc_gesi}
\title{captura os termos GESI}
\usage{
dtc_gesi(df, processo, motor)
}
\arguments{
\item{df}{dataset que contém os projetos de energia}

\item{processo}{atributo de codigo único do projeto}

\item{motor}{vetor com os atributos utilizados para a classificação}
}
\value{

}
\description{
captura os termos GESI
}
\examples{
dtc_gesi(dataset, id, motor)
}
