% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclui_projetos.R
\name{exclui_projetos}
\alias{exclui_projetos}
\title{Exclui projetos do sqlite}
\usage{
exclui_projetos(fonte, sqlite)
}
\arguments{
\item{fonte}{string de caracteres da  fonte dos projetos que desejam ser removidos}

\item{sqlite}{diretório do sqlite}
}
\value{

}
\description{
Exclui projetos do sqlite
}
