% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executa_carga_completa.R
\name{executa_carga_completa}
\alias{executa_carga_completa}
\title{executa a carga completa no SQLite, removendo todos os casos da fonte inserida e inserindo todos os casos do dataset}
\usage{
executa_carga_completa(df, sqlite)
}
\arguments{
\item{df}{dataset tratado e validado}

\item{sqlite}{diretório do sqlite}
}
\value{

}
\description{
executa a carga completa no SQLite, removendo todos os casos da fonte inserida e inserindo todos os casos do dataset
}
\examples{
executa_carga_completa(df_validado_aneel, diretorio_sqlite)
executa_carga_completa(df_validado_anp, diretorio_sqlite)
executa_carga_completa(df_validado_bndes, diretorio_sqlite)
executa_carga_completa(df_validado_cnen, diretorio_sqlite)

}
