% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executa_carga_gc.R
\name{executa_carga_gc}
\alias{executa_carga_gc}
\title{Executa a carga incremental alterando os valores do projeto agregado, inserindo casos novos e atualizando atributos}
\usage{
executa_carga_gc(df, sqlite)
}
\arguments{
\item{df}{dataset validado a ser inserido no sqlite}

\item{sqlite}{diretório do sqlite}
}
\value{

}
\description{
Executa a carga incremental alterando os valores do projeto agregado, inserindo casos novos e atualizando atributos
}
\examples{
executa_carga_gc(df_validado_aneel, diretorio_sqlite)
executa_carga_gc(df_validado_anp, diretorio_sqlite)
executa_carga_gc(df_validado_bndes, diretorio_sqlite)
executa_carga_gc(df_validado_cnen, diretorio_sqlite)
}
