% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executa_carga_incremental.R
\name{executa_carga_incremental}
\alias{executa_carga_incremental}
\title{executa carga incremental no SQLite, adicionando apenas os casos que não foram identificados no SQlite e atualizando os atributos}
\usage{
executa_carga_incremental(df, sqlite)
}
\arguments{
\item{df}{dataset validado}

\item{sqlite}{diretório do sqlite}
}
\value{

}
\description{
executa carga incremental no SQLite, adicionando apenas os casos que não foram identificados no SQlite e atualizando os atributos
}
\examples{
executa_carga_incremental(df_validado_aneel, diretorio_sqlite)
executa_carga_incremental(df_validado_anp, diretorio_sqlite)
executa_carga_incremental(df_validado_bndes, diretorio_sqlite)
executa_carga_incremental(df_validado_cnen, diretorio_sqlite)
}
