% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executa_tratamento_completo.R
\name{executa_tratamento_completo}
\alias{executa_tratamento_completo}
\title{executa o tratamento que gera um dataset para a carga completa do sqlite}
\usage{
executa_tratamento_completo(df, sqlite)
}
\arguments{
\item{df}{dataset tratado}

\item{sqlite}{diretório do .db do sqlite}
}
\value{

}
\description{
executa o tratamento que gera um dataset para a carga completa do sqlite
}
\examples{
executa_tratamento_completo(df_intermediario_aneel, diretorio_sqlite)
executa_tratamento_completo(df_intermediario_anp, diretorio_sqlite)
executa_tratamento_completo(df_intermediario_bndes, diretorio_sqlite)
executa_tratamento_completo(df_intermediario_cnen, diretorio_sqlite)
}
