% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_proj_gc.R
\name{fun_proj_gc}
\alias{fun_proj_gc}
\title{faz o calculo do projeto guarda chuva.}
\usage{
fun_proj_gc(base, ft_guarda_chuva)
}
\arguments{
\item{base}{dataframe que contem o projeto de energia}

\item{ft_guarda_chuva}{subconjunto dos projeto que agregam os valores de financiamento do projeto}
}
\value{

}
\description{
faz o calculo do projeto guarda chuva.
}
\examples{
fun_proj_gc(dataset, projeto_guarda_chuva)
}
