% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valida_dataset2.R
\name{valida_dataset2}
\alias{valida_dataset2}
\title{faz a validação dos casos para o script}
\usage{
valida_dataset2(df_original, df_inserir)
}
\arguments{
\item{df_inserir}{}
}
\value{

}
\description{
faz a validação dos casos para o script
}
