% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valida_termos_anp.R
\name{valida_termos_anp}
\alias{valida_termos_anp}
\title{reclassifica em cima de alguns parametros a classificação da função dtc_categorias}
\usage{
valida_termos_anp(df, categorias)
}
\arguments{
\item{df}{dataframe que contem os projetos de energia}

\item{categorias}{atributo que mede as classificações das categorias de IEA}
}
\value{

}
\description{
reclassifica em cima de alguns parametros a classificação da função dtc_categorias
}
\examples{
valida_termos_anp(dataset, categorias)
}
