% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_base_intermediaria_aneel.R
\name{cria_base_intermediaria_aneel}
\alias{cria_base_intermediaria_aneel}
\title{Cria a base intemediária para a aneel}
\usage{
cria_base_intermediaria_aneel(
  origem_processos = here::here("data/SGPED_BI/PD Busca Textual.csv"),
  origem_equipes = here::here("data/SGPED_BI/5.PD RF EQUIPE.csv")
)
}
\value{

}
\description{
Cria a base intemediária para a aneel
}
