% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_base_intermediaria_cnen.R
\name{cria_base_intermediaria_cnen}
\alias{cria_base_intermediaria_cnen}
\title{Cria a base intemediária para o CNEN}
\usage{
cria_base_intermediaria_cnen(
  origem_processos = here::here("data/CNEN/Projeto CNEN_Plataforma Inova-E.xlsx")
)
}
\value{

}
\description{
Cria a base intemediária para o CNEN
}
